# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Work Stat Lite Api Documentation",
        "description": "Work Stat Lite Api Documentation",
        "contact": {
            "email": "orutu1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://wslite-backend.herokuapp.com/api//",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/holidays": {
            "get": {
                "tags": [
                    "Holidays"
                ],
                "summary": "Get list of all Holidays",
                "description": "Returns list of Holidays",
                "operationId": "GetHolidays",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/holiday-create": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Holiday create",
                "description": "Holiday create",
                "operationId": "postSchedule",
                "parameters": [
                    {
                        "name": "Name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "date",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Holiday created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/organizations": {
            "get": {
                "tags": [
                    "organizations"
                ],
                "summary": "Get list of all organizations",
                "description": "Returns list of organizations",
                "operationId": "GetOrganizations",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/organization-create": {
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "Organization create",
                "description": "Organization create",
                "operationId": "postOrganization",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true
                    },
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "organization domain",
                        "required": true
                    },
                    {
                        "name": "image_path",
                        "in": "query",
                        "description": "office logo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/organization-show": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get organization details",
                "description": "Returns organization details",
                "operationId": "getOrganizationDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Json"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/organization-update": {
            "patch": {
                "tags": [
                    "Organization"
                ],
                "summary": "Update Organization records",
                "description": "Update Organization records",
                "operationId": "updateOrganization",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email"
                    },
                    {
                        "name": "domain",
                        "in": "query",
                        "description": "organization domain"
                    },
                    {
                        "name": "image_path",
                        "in": "query",
                        "description": "office logo"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/organization-delete": {
            "delete": {
                "tags": [
                    "Organization"
                ],
                "summary": "Delete organization records",
                "description": "Delete organization records",
                "operationId": "deleteOrganization",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/schedule-create": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Register",
                "description": "Register Schedule",
                "operationId": "postSchedule",
                "parameters": [
                    {
                        "name": "Title",
                        "in": "query",
                        "description": "Title",
                        "required": true
                    },
                    {
                        "name": "week_days",
                        "in": "query",
                        "description": "Week Days Json",
                        "required": true
                    },
                    {
                        "name": "schedule_in",
                        "in": "query",
                        "description": "Schedule start time",
                        "required": true
                    },
                    {
                        "name": "schedule_out",
                        "in": "query",
                        "description": "Schedule work time",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
...
        "/schedule-update": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Register",
                "description": "Update Schedule",
                "operationId": "postSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    },
                    {
                        "name": "Title",
                        "in": "query",
                        "description": "Title"
                    },
                    {
                        "name": "week_days",
                        "in": "query",
                        "description": "Week Days Json"
                    },
                    {
                        "name": "schedule_in",
                        "in": "query",
                        "description": "Schedule start time"
                    },
                    {
                        "name": "schedule_out",
                        "in": "query",
                        "description": "Schedule work time"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule updated successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}